fix: renamed assets folder and fixed tilt effect
