+++ package.json
@@ -24,10 +24,10 @@
     "react-dom": "^17.0.2",
     "react-github-calendar": "^3.2.2",
     "react-icons": "^4.8.0",
-    "react-parallax-tilt": "^1.7.277",
     "react-pdf": "^5.7.1",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
+    "react-tilt": "^1.0.2",
     "react-tsparticles": "^1.42.2",
     "typewriter-effect": "^2.18.2",
     "web-vitals": "^2.1.4"
@@ -51,6 +51,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-import": "^2.31.0",
