+++ src/style.css
@@ -33,7 +33,7 @@ button:focus {
   height: 100%;
   z-index: 999999;
   background-color: #0c0513;
-  background-image: url(./assets/pre.svg);
+  background-image: url(./temp-assets/pre.svg);
   background-repeat: no-repeat;
   background-position: center;
 }
@@ -257,7 +257,7 @@ button:focus {
 .home-section {
   position: relative;
   z-index: -1;
-  background-image: var(--image-gradient), url(./assets/home-bg.jpg);
+  background-image: var(--image-gradient), url(./temp-assets/home-bg.jpg);
   background-position: top center;
   background-repeat: no-repeat;
   padding-bottom: 30px !important;
