+++ src/temp-assets/Projects/emotion.png
