+++ src/temp-assets/rajat_singh_avatar.png
