+++ src/temp-assets/rajat_singh_circle.png
